Fix project package
