+++ app/build.gradle
@@ -9,7 +9,7 @@ android {
     buildToolsVersion "23.0.1"
 
     defaultConfig {
-        applicationId "redhat.com.syncsample"
+        applicationId "com.feedhenry.sync"
         minSdkVersion 10
         targetSdkVersion 23
         versionCode 1
