+++ app/src/main/java/com/feedhenry/sync/item/ShoppingItem.java
@@ -4,7 +4,7 @@
  * Please refer to your contract with FeedHenry for the software license agreement.
  * If you do not have a contract, you do not have a license to use this software.
  */
-package redhat.com.syncsample.item;
+package com.feedhenry.sync.item;
 
 import com.feedhenry.sdk.sync.FHSyncUtils;
