+++ app/src/main/java/com/feedhenry/sync/view/ListItemsFragment.java
@@ -4,7 +4,7 @@
  * Please refer to your contract with FeedHenry for the software license agreement.
  * If you do not have a contract, you do not have a license to use this software.
  */
-package redhat.com.syncsample.view;
+package com.feedhenry.sync.view;
 
 import android.content.DialogInterface;
 import android.support.design.widget.FloatingActionButton;
@@ -33,11 +33,11 @@ import org.json.fh.JSONObject;
 import java.util.Iterator;
 import java.util.TreeSet;
 
-import redhat.com.syncsample.R;
-import redhat.com.syncsample.item.ShoppingItem;
-import redhat.com.syncsample.item.ShoppingItemAdapter;
-import redhat.com.syncsample.item.ShoppingItemSelectHandler;
-import redhat.com.syncsample.item.gesturehelper.UncoverDeleteGestureCallback;
+import com.feedhenry.sync.R;
+import com.feedhenry.sync.item.ShoppingItem;
+import com.feedhenry.sync.item.ShoppingItemAdapter;
+import com.feedhenry.sync.item.ShoppingItemSelectHandler;
+import com.feedhenry.sync.item.gesturehelper.UncoverDeleteGestureCallback;
 
 /**
  * This class sets up synchronization, displays Shopping Items, and sends events to the sync system.
