+++ app/src/main/res/layout/content_main.xml
@@ -9,7 +9,7 @@
 <fragment xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
     xmlns:app="http://schemas.android.com/apk/res-auto" android:id="@+id/fragment"
-    android:name="redhat.com.syncsample.view.ListItemsFragment" android:layout_width="match_parent"
+    android:name="com.feedhenry.sync.view.ListItemsFragment" android:layout_width="match_parent"
     android:layout_height="match_parent"
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
     tools:layout="@layout/fragment_main" />
