+++ app/src/main/res/layout/fragment_main.xml
@@ -11,7 +11,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="redhat.com.syncsample.view.ListItemsFragment"
+    tools:context="com.feedhenry.sync.view.ListItemsFragment"
     tools:showIn="@layout/chrome_main">
 
     <android.support.v7.widget.RecyclerView
